[sgen] Don't crash on nurseries of 512M or larger.
Commit migrated from https://github.com/mono/mono/commit/5db0301dc89a51b8d7bd970e1a1a552088596499

+++ src/mono/mono/metadata/sgen-gc.h
@@ -143,7 +143,7 @@ struct _GCMemSection {
 	/* in major collections indexes in the pin_queue for objects that pin this section */
 	void **pin_queue_start;
 	int pin_queue_num_entries;
-	unsigned short num_scan_start;
+	unsigned int num_scan_start;
 };
 
 /*
